chore(types): modularize base types
