+++ packages/background/src/lib/context-menu.ts
@@ -2,7 +2,7 @@ import { v4 as uuidv4 } from "uuid";
 
 import { browser, popoutExtension } from "@worm/shared/src/browser";
 import { storageGetByKeys, storageSetByKeys } from "@worm/shared/src/storage";
-import { QueryPattern } from "@worm/types";
+import { QueryPattern } from "@worm/types/src/replace";
 
 const ADD_NEW_RULE_ID = "add-new-rule";
