+++ packages/popup/cypress/support/commands.ts
@@ -1,8 +1,9 @@
 /// <reference types="cypress" />
 import "@testing-library/cypress/add-commands";
+import "./cmd";
 
 import { TestBrowser } from "@worm/testing";
-import { Storage } from "@worm/types";
+import { Storage } from "@worm/types/src/storage";
 
 import { VisitWithStorageParams } from "./types";
 
@@ -18,5 +19,3 @@ declare global {
     }
   }
 }
-
-import "./cmd";
