+++ packages/popup/cypress/support/types.ts
@@ -1,3 +1,4 @@
-import { DeepPartial, StorageProvider } from "@worm/types";
+import { DeepPartial } from "@worm/types";
+import { StorageProvider } from "@worm/types/src/storage";
 
 export type VisitWithStorageParams = DeepPartial<Record<StorageProvider, any>>;
