+++ packages/shared/src/replace/lib/suggest.ts
@@ -1,4 +1,5 @@
-import { ReplacementSuggest, ToneOption } from "@worm/types";
+import { ToneOption } from "@worm/types/src/api";
+import { ReplacementSuggest } from "@worm/types/src/replace";
 
 export const DEFAULT_TONE_OPTION: ToneOption = "casual";
