+++ packages/testing/src/test-browser.ts
@@ -1,5 +1,5 @@
 import { DEFAULT_REPLACEMENT_STYLE } from "@worm/shared/src/replace/lib/style";
-import { Storage } from "@worm/types";
+import { Storage } from "@worm/types/src/storage";
 
 import MockBrowser from "./mock-webextension-polyfill";
